requirements.txt
curl_cffi
httpx
python-dotenv
fastapi 
uvicorn
